refactor the main ss file
