--- policyengine_us/parameters/gov/states/ne/tax/income/agi/subtractions/social_security/in_effect.yaml
@@ -1,12 +1,0 @@
-description: Nebraska provides this Federal Retirement Income Exclusion on agi if this is true. 
-values: 
-  2024-01-01: true
-metadata:
-  unit: bool
-  period: year
-  label: Nebraska AGI federal retirement income exclusion
-  reference:
-   - title: 2023 Nebraska Legislative Changes, Federal Retirement Income Exclusion
-     href: https://revenue.nebraska.gov/about/2023-nebraska-legislative-changes
-   - title: Section 77-2716(20)
-     href: https://www.nebraskalegislature.gov/FloorDocs/108/PDF/Slip/LB754.pdf#page=10
