+++ policyengine_us/parameters/gov/states/ne/tax/income/agi/subtractions/subtractions.yaml
@@ -1,0 +1,17 @@
+description: Nebraska subtracts these sources from adjusted gross income.
+values:
+  2021-01-01:
+    - ne_social_security_subtraction
+  2024-01-01:
+    - ne_social_security_subtraction
+    - taxable_public_pension_income
+    
+
+metadata:
+  unit: list
+  period: year
+  label: Nebraska subtractions from federal adjusted gross income
+  reference:
+  - title: 2021 NE income tax form and instruction booklet
+    href: https://revenue.nebraska.gov/files/doc/tax-forms/2021/f_1040n_booklet.pdf
+
