+++ policyengine_us/tests/policy/baseline/gov/states/ne/tax/income/adjusted_gross_income/ne_social_security_subtraction.yaml
@@ -23,7 +23,7 @@
         members: [person1, person2]
         state_code: NE
   output:
-    ne_agi_subtractions: 10_000 + 10_000
+    ne_social_security_subtraction: 10_000 + 10_000
 
 - name: NE AGI subtractions unit test 2
   absolute_error_margin: 0.01
@@ -34,7 +34,7 @@
      taxable_social_security: 10_000
      state_code: NE
   output:
-    ne_agi_subtractions: 500
+    ne_social_security_subtraction: 500
 
 - name: NE AGI subtractions unit test 3
   absolute_error_margin: 0.01
@@ -45,4 +45,4 @@
      taxable_social_security: 10_000
      state_code: NE
   output:
-    ne_agi_subtractions: 10_000
+    ne_social_security_subtraction: 10_000
