+++ policyengine_us/variables/gov/states/ne/tax/income/adjusted_gross_income/ne_agi_subtractions.py
@@ -4,7 +4,7 @@ from policyengine_us.model_api import *
 class ne_agi_subtractions(Variable):
     value_type = float
     entity = TaxUnit
-    label = "NE AGI subtractions from federal AGI"
+    label = "Nebraska subtractions from federal adjusted gross income"
     unit = USD
     definition_period = YEAR
     reference = (
@@ -15,21 +15,4 @@ class ne_agi_subtractions(Variable):
     )
     defined_for = StateCode.NE
 
-    def formula(tax_unit, period, parameters):
-        fagi = tax_unit("adjusted_gross_income", period)
-        filing_status = tax_unit("filing_status", period)
-        p = parameters(period).gov.states.ne.tax.income.agi.subtractions
-        oasdi_fraction = where(
-            fagi <= p.social_security.threshold[filing_status],
-            1.0,
-            p.social_security.fraction,
-        )
-        taxable_oasdi = add(tax_unit, period, ["taxable_social_security"])
-        oasdi_amount = taxable_oasdi * oasdi_fraction
-        if p.in_effect:
-            taxable_retirement_pension = add(
-                tax_unit, period, ["taxable_public_pension_income"]
-            )
-            return taxable_retirement_pension + oasdi_amount
-        else:
-            return oasdi_amount
+    adds = "gov.states.ne.tax.income.agi.subtractions.subtractions"
